prototype/requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
Django==4.0.3
Flask==2.1.1
Flask-Cors==3.0.10
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
python-dotenv==0.20.0
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
Werkzeug==2.1.1
zipp==3.8.0
